% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gestage.R
\docType{data}
\name{coefs_mod_c}
\alias{coefs_mod_c}
\title{Coefficients to approximate Model C fits}
\format{
\subsection{\code{coefs_mod_c}}{

A matrix
}
}
\usage{
coefs_mod_c
}
\description{
Coefficients to approximate Model C fits
}
\keyword{datasets}
