% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gestage.R
\docType{data}
\name{coefs_mod_d}
\alias{coefs_mod_d}
\title{Coefficients to approximate Model D fits}
\format{
\subsection{\code{coefs_mod_d}}{

A matrix
}
}
\usage{
coefs_mod_d
}
\description{
Built from spline interpolation of actual model predicted
values
}
\keyword{datasets}
