% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gestage_c}
\alias{gestage_c}
\title{Predict gestational age using Model C}
\usage{
gestage_c(gagelmp, birthwt, round_pred = TRUE)
}
\arguments{
\item{gagelmp}{Gestational age at birth in days based on last menstrual
period (must be between 161 and 350).}

\item{birthwt}{Birth weight in grams (must be between 1000 and 5000).}

\item{round_pred}{Should the predicted gestational age be rounded to the
nearest day? Default is \code{TRUE}.}
}
\description{
Predict gestational age using Model C
}
\examples{
gestage_c(gagelmp = 250, birthwt = 2800)
gestage_c(gagelmp = 201:225, birthwt = 2800)
gestage_c(gagelmp = 250, birthwt = seq(3005, 3100, by = 5))
}
\references{
Simplified models to assess newborn gestational age in low-middle
income countries: findings from a multicountry, prospective cohort study.
\emph{BMJ Global Health}, 2021.
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8438948/}{(url)}
}
