% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gestage_d}
\alias{gestage_d}
\title{Predict gestational age using Model C}
\usage{
gestage_d(birthhc, birthwt, round_pred = TRUE)
}
\arguments{
\item{birthhc}{birth head circumference in centimeters (must be between
25 and 40).}

\item{birthwt}{birth weight in grams (must be between 1000 and 5000).}

\item{round_pred}{Should the predicted gestational age be rounded to the
nearest day? Default is \code{TRUE}.}
}
\description{
Predict gestational age using Model C
}
\examples{
gestage_d(birthhc = 35, birthwt = 2800)
gestage_d(birthhc = seq(25, 30, by = 0.1), birthwt = 2800)
gestage_d(birthhc = 35, birthwt = seq(3005, 3100, by = 5))
}
\references{
Simplified models to assess newborn gestational age in low-middle
income countries: findings from a multicountry, prospective cohort study.
\emph{BMJ Global Health}, 2021.
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8438948/}{(url)}
}
