% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_c.R
\name{method_c_predict}
\alias{method_c_predict}
\title{Predict gestational age using Model C}
\usage{
method_c_predict(gagelmp, birthwt)
}
\arguments{
\item{gagelmp}{gestational age at birth in days based on last menstrual
period (must be between 161 and 350)}

\item{birthwt}{birth weight in grams (must be between 1000 and 5000)}
}
\description{
Predict gestational age using Model C
}
\examples{
method_c_predict(gagelmp = 250, birthwt = 2800)
method_c_predict(gagelmp = 201:225, birthwt = 2800)
method_c_predict(gagelmp = 250, birthwt = seq(3005, 3100, by = 5))
}
\references{
Simplified models to assess newborn gestational age in low-middle
income countries: findings from a multicountry, prospective cohort study.
\emph{BMJ Global Health}, 2021.
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8438948/}{(url)}
}
